split up tests into fixable and unfixable now and add annotations
